pyproject.toml
[tool.poetry]
name = "brettspill_py"
version = "0.1.0"
description = "Brettpill på norsk"
authors = ["Espen Holtebu <espen.holtebu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.0"
djangorestframework = "^3.13.0"
psycopg2 = "^2.9.1"
django-environ = "^0.8.1"
gunicorn = "^20.1.0"
django-colorfield = "^0.7.2"
whitenoise = "^6.2.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"
django-extensions = "^3.2.1"
drf-flex-fields = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"
pytest-django = "^4.5.2"
pytest-dotenv = "^0.5.2"
pytest-env = "^0.6.2"


[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"
